requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
openai==1.30.5
chromadb==0.4.24
pypdf==4.2.0
python-multipart==0.0.9
pydantic==1.10.15
pydantic-settings==2.2.1
ldap3==2.9.1
google-auth==2.29.0
google-auth-httplib2==0.2.0
